+++ bm/InterfaceHelpers.swift
@@ -80,3 +80,13 @@ extension UIColor {
         return UIColor(named: "BMRed")!
     }
 }
+
+extension String {
+    func capitalizingFirstLetter() -> String {
+        return prefix(1).capitalized + dropFirst()
+    }
+
+    mutating func capitalizeFirstLetter() {
+        self = self.capitalizingFirstLetter()
+    }
+}