+++ bm/Agenda/AgendaParser.swift
@@ -87,7 +87,7 @@ class AgendaParser {
         }
 
         let category = infoStringComponents[0].trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
-        let library = infoStringComponents[1].components(separatedBy: "-").first?.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
+        let library = infoStringComponents[1].components(separatedBy: "  -").first?.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
 
         let image: URL?
         if let imageString = html.parse(between: "<img src=\"", and: "\""),